requirements.txt
langchain[openai]
openpyxl
pandas
python-dotenv
